app/requirements.txt
Flask==2.0.1
flask_sqlalchemy==2.5.1
lxml==4.6.3
numpy
pandas
pyyaml
requests==2.26.0
json2html==1.3.0
psycopg-binary
psycopg2==2.8.4
sklearn
tensorflow
gunicorn
